File comparison added.
While it still adds the number between the ESSID and the "clean" word, now checks for duplicates and eventually removes them before passing them through wpaclean.
The unique caps are stored in the B/UniqueCaps folder.

The comparison is made using the md5sum program (included in most Linux distribution).